Added a query logger. Will begin working on gradually switching to POST eventually.

+++ src/com/lightcraftmc/database/util/UtilCategoriesPage.java
@@ -3,13 +3,27 @@ package com.lightcraftmc.database.util;
 import java.io.File;
 import java.util.ArrayList;
 
+import com.lightcraftmc.database.command.WebCommandInterpreter;
+
 public class UtilCategoriesPage {
 
-    public static ArrayList<String> getCategoriesPage(String[] args) {
-        if (args.length == 1) {
-            return categoriesMainPage(Banner.blank());
-        }
+    public static ArrayList<String> getCategoriesPage(String[] args, String ip) {
         Banner banner = Banner.blank();
+        if (args[0].contains("!")) {
+            String create = args[0].split("!")[1];
+            WebCommandInterpreter.getInstance().interpret("insert " + create + " temp Created_by_web_interface", ip, true);
+            RawCategory r = getCategory(create);
+            for (File f : r.getItems()) {
+                if (!f.isDirectory()) {
+                    f.delete();
+                }
+            }
+            banner = Banner.defaultSuccess();
+            args[0] = args[0].split("!")[0];
+        }
+        if (args.length == 1) {
+            return categoriesMainPage(banner);
+        }
         boolean categoriesPage = false;
         if (args.length > 2) {
             String command = args[2];
@@ -26,8 +40,11 @@ public class UtilCategoriesPage {
             if (command.equalsIgnoreCase("empty")) {
                 try {
                     RawCategory r = getCategory(args[1]);
-                    for (File f : r.getItems())
-                        f.delete();
+                    for (File f : r.getItems()) {
+                        if (!f.isDirectory()) {
+                            f.delete();
+                        }
+                    }
                     banner = Banner.defaultSuccess();
                 } catch (Exception ex) {
                     banner = Banner.defaultFailed();
@@ -52,8 +69,15 @@ public class UtilCategoriesPage {
         lines.addAll(UtilBootstrap.containerOpen());
         lines.addAll(UtilBootstrap.generateTabs("Categories"));
         lines.addAll(UtilBootstrap.containerClose());
-
         lines.addAll(UtilBootstrap.containerOpen());
+        lines.add("<br>");
+        lines.add("<form class=\"well span9\" method=\"GET\" action=\"/?categories!\">");
+        lines.add("<div class=\"input-group input-group-lg\">");
+        lines.add("<span class=\"input-group-addon\" id=\"sizing-addon1\">Create Category</span>");
+        lines.add("<input type=\"text\" class=\"form-control\" placeholder=\"Title\" name=\"createcategory\" aria-describedby=\"sizing-addon1\">");
+        lines.add("</div>");
+        lines.add("</form>");
+        lines.add("<br>");
         lines.addAll(UtilBootstrap.generateCategoryListing("Available Categories", UtilGenerateCategories.getCategories(), "/?categories-"));
         lines.addAll(UtilBootstrap.containerClose());
         lines.add(Tag.close("body"));
@@ -80,7 +104,7 @@ public class UtilCategoriesPage {
         lines.addAll(UtilBootstrap.createSmallButton("< Go back", "/?categories", "info"));
         ArrayList<Link> dropdown = new ArrayList<Link>();
         ArrayList<Link> dropdown2 = new ArrayList<Link>();
-        dropdown.add(new Link(pfx + "allowpublic", "Allow Public Access"));
+        // dropdown.add(new Link(pfx + "allowpublic", "Allow Public Access"));
         dropdown2.add(new Link(pfx, "Are you sure?"));
         dropdown2.add(new Link(pfx, "<b>This will also remove sub-categories!</b>"));
         dropdown2.add(Link.seperator());
@@ -92,9 +116,16 @@ public class UtilCategoriesPage {
         lines.addAll(UtilBootstrap.generateDropdown("Delete", ButtonType.DANGER, dropdown2));
         lines.addAll(UtilBootstrap.containerClose());
         lines.add("<br>");
+        try{
         lines.addAll(UtilBootstrap.containerOpen());
         lines.addAll(UtilBootstrap.generateTable(category));
         lines.addAll(UtilBootstrap.containerClose());
+        lines.addAll(UtilBootstrap.containerOpen());
+        lines.addAll(UtilBootstrap.generateSubcategoriesTable(category));
+        lines.addAll(UtilBootstrap.containerClose());
+        }catch(Exception ex){
+            ex.printStackTrace();
+        }
         lines.add(Tag.close("body"));
         lines.add(Tag.close("html"));
         return lines;
